[#157903353] Create Welcome component
