+++ client/assets/css/index.css
@@ -67,8 +67,6 @@
 
 	}
 
-	
-
 } 
 
 @media only screen and (min-width: 768px) {
