+++ client/src/components/Home/home.scss
@@ -7,5 +7,5 @@
 }
 
 .hungry.bounceIn{
-    font-size: 40px;
+  font-size: 40px;
 }
